Separate the MSP handling to elrs.c in core
